Formats solidity using prettier

+++ contracts/external/chai/chai.sol
@@ -23,63 +23,79 @@ interface VatLike {
 
 interface PotLike {
     function chi() external returns (uint256);
+
     function rho() external returns (uint256);
+
     function drip() external returns (uint256);
+
     function join(uint256) external;
+
     function exit(uint256) external;
 }
 
 interface JoinLike {
-    function join(address, uint) external;
-    function exit(address, uint) external;
+    function join(address, uint256) external;
+
+    function exit(address, uint256) external;
 }
 
 interface GemLike {
-    function transferFrom(address,address,uint) external returns (bool);
-    function approve(address,uint) external returns (bool);
+    function transferFrom(
+        address,
+        address,
+        uint256
+    ) external returns (bool);
+
+    function approve(address, uint256) external returns (bool);
 }
 
 contract Chai {
     // --- Data ---
-    VatLike  public vat;
-    PotLike  public pot;
+    VatLike public vat;
+    PotLike public pot;
     JoinLike public daiJoin;
-    GemLike  public daiToken;
+    GemLike public daiToken;
 
     // --- ERC20 Data ---
-    string  public constant name     = "Chai";
-    string  public constant symbol   = "CHAI";
-    string  public constant version  = "1";
-    uint8   public constant decimals = 18;
+    string public constant name = 'Chai';
+    string public constant symbol = 'CHAI';
+    string public constant version = '1';
+    uint8 public constant decimals = 18;
     uint256 public totalSupply;
 
-    mapping (address => uint)                      public balanceOf;
-    mapping (address => mapping (address => uint)) public allowance;
-    mapping (address => uint)                      public nonces;
+    mapping(address => uint256) public balanceOf;
+    mapping(address => mapping(address => uint256)) public allowance;
+    mapping(address => uint256) public nonces;
 
-    event Approval(address indexed src, address indexed guy, uint wad);
-    event Transfer(address indexed src, address indexed dst, uint wad);
+    event Approval(address indexed src, address indexed guy, uint256 wad);
+    event Transfer(address indexed src, address indexed dst, uint256 wad);
 
     // --- Math ---
-    uint constant RAY = 10 ** 27;
-    function add(uint x, uint y) internal pure returns (uint z) {
+    uint256 constant RAY = 10**27;
+
+    function add(uint256 x, uint256 y) internal pure returns (uint256 z) {
         require((z = x + y) >= x);
     }
-    function sub(uint x, uint y) internal pure returns (uint z) {
+
+    function sub(uint256 x, uint256 y) internal pure returns (uint256 z) {
         require((z = x - y) <= x);
     }
-    function mul(uint x, uint y) internal pure returns (uint z) {
+
+    function mul(uint256 x, uint256 y) internal pure returns (uint256 z) {
         require(y == 0 || (z = x * y) / y == x);
     }
-    function rmul(uint x, uint y) internal pure returns (uint z) {
+
+    function rmul(uint256 x, uint256 y) internal pure returns (uint256 z) {
         // always rounds down
         z = mul(x, y) / RAY;
     }
-    function rdiv(uint x, uint y) internal pure returns (uint z) {
+
+    function rdiv(uint256 x, uint256 y) internal pure returns (uint256 z) {
         // always rounds down
         z = mul(x, RAY) / y;
     }
-    function rdivup(uint x, uint y) internal pure returns (uint z) {
+
+    function rdivup(uint256 x, uint256 y) internal pure returns (uint256 z) {
         // always rounds up
         z = add(mul(x, RAY), sub(y, 1)) / y;
     }
@@ -87,9 +103,14 @@ contract Chai {
     // --- EIP712 niceties ---
     bytes32 public constant DOMAIN_SEPARATOR = 0x0b50407de9fa158c2cba01a99633329490dfd22989a150c20e8c7b4c1fb0fcc3;
     // keccak256("Permit(address holder,address spender,uint256 nonce,uint256 expiry,bool allowed)"));
-    bytes32 public constant PERMIT_TYPEHASH  = 0xea2aa0a1be11a07ed86d755c93467f4f82362b452371d1ba94d1715123511acb;
+    bytes32 public constant PERMIT_TYPEHASH = 0xea2aa0a1be11a07ed86d755c93467f4f82362b452371d1ba94d1715123511acb;
 
-    constructor(address vat_, address pot_, address daiJoin_, address weth_) public {
+    constructor(
+        address vat_,
+        address pot_,
+        address daiJoin_,
+        address weth_
+    ) public {
         /* assert (DOMAIN_SEPARATOR ==
           keccak256(abi.encode(
             keccak256("EIP712Domain(string name,string version,uint256 chainId,address verifyingContract)"),
@@ -104,25 +125,33 @@ contract Chai {
         vat.hope(address(daiJoin));
         vat.hope(address(pot));
 
-        daiToken.approve(address(daiJoin), uint(-1));
+        daiToken.approve(address(daiJoin), uint256(-1));
     }
 
     // --- Token ---
-    function transfer(address dst, uint wad) external returns (bool) {
+    function transfer(address dst, uint256 wad) external returns (bool) {
         return transferFrom(msg.sender, dst, wad);
     }
+
     // like transferFrom but dai-denominated
-    function move(address src, address dst, uint wad) external returns (bool) {
-        uint chi = (now > pot.rho()) ? pot.drip() : pot.chi();
+    function move(
+        address src,
+        address dst,
+        uint256 wad
+    ) external returns (bool) {
+        uint256 chi = (now > pot.rho()) ? pot.drip() : pot.chi();
         // rounding up ensures dst gets at least wad dai
         return transferFrom(src, dst, rdivup(wad, chi));
     }
-    function transferFrom(address src, address dst, uint wad)
-        public returns (bool)
-    {
-        require(balanceOf[src] >= wad, "chai/insufficient-balance");
-        if (src != msg.sender && allowance[src][msg.sender] != uint(-1)) {
-            require(allowance[src][msg.sender] >= wad, "chai/insufficient-allowance");
+
+    function transferFrom(
+        address src,
+        address dst,
+        uint256 wad
+    ) public returns (bool) {
+        require(balanceOf[src] >= wad, 'chai/insufficient-balance');
+        if (src != msg.sender && allowance[src][msg.sender] != uint256(-1)) {
+            require(allowance[src][msg.sender] >= wad, 'chai/insufficient-allowance');
             allowance[src][msg.sender] = sub(allowance[src][msg.sender], wad);
         }
         balanceOf[src] = sub(balanceOf[src], wad);
@@ -130,44 +159,53 @@ contract Chai {
         emit Transfer(src, dst, wad);
         return true;
     }
-    function approve(address usr, uint wad) external returns (bool) {
+
+    function approve(address usr, uint256 wad) external returns (bool) {
         allowance[msg.sender][usr] = wad;
         emit Approval(msg.sender, usr, wad);
         return true;
     }
 
     // --- Approve by signature ---
-    function permit(address holder, address spender, uint256 nonce, uint256 expiry, bool allowed, uint8 v, bytes32 r, bytes32 s) external
-    {
-        bytes32 digest = keccak256(abi.encodePacked(
-            "\x19\x01",
-            DOMAIN_SEPARATOR,
-            keccak256(abi.encode(PERMIT_TYPEHASH,
-                                 holder,
-                                 spender,
-                                 nonce,
-                                 expiry,
-                                 allowed))));
-        require(holder != address(0), "chai/invalid holder");
-        require(holder == ecrecover(digest, v, r, s), "chai/invalid-permit");
-        require(expiry == 0 || now <= expiry, "chai/permit-expired");
-        require(nonce == nonces[holder]++, "chai/invalid-nonce");
+    function permit(
+        address holder,
+        address spender,
+        uint256 nonce,
+        uint256 expiry,
+        bool allowed,
+        uint8 v,
+        bytes32 r,
+        bytes32 s
+    ) external {
+        bytes32 digest =
+            keccak256(
+                abi.encodePacked(
+                    '\x19\x01',
+                    DOMAIN_SEPARATOR,
+                    keccak256(abi.encode(PERMIT_TYPEHASH, holder, spender, nonce, expiry, allowed))
+                )
+            );
+        require(holder != address(0), 'chai/invalid holder');
+        require(holder == ecrecover(digest, v, r, s), 'chai/invalid-permit');
+        require(expiry == 0 || now <= expiry, 'chai/permit-expired');
+        require(nonce == nonces[holder]++, 'chai/invalid-nonce');
 
-        uint can = allowed ? uint(-1) : 0;
+        uint256 can = allowed ? uint256(-1) : 0;
         allowance[holder][spender] = can;
         emit Approval(holder, spender, can);
     }
 
-    function dai(address usr) external returns (uint wad) {
-        uint chi = (now > pot.rho()) ? pot.drip() : pot.chi();
+    function dai(address usr) external returns (uint256 wad) {
+        uint256 chi = (now > pot.rho()) ? pot.drip() : pot.chi();
         wad = rmul(chi, balanceOf[usr]);
     }
+
     // wad is denominated in dai
-    function join(address dst, uint wad) external {
-        uint chi = (now > pot.rho()) ? pot.drip() : pot.chi();
-        uint pie = rdiv(wad, chi);
+    function join(address dst, uint256 wad) external {
+        uint256 chi = (now > pot.rho()) ? pot.drip() : pot.chi();
+        uint256 pie = rdiv(wad, chi);
         balanceOf[dst] = add(balanceOf[dst], pie);
-        totalSupply    = add(totalSupply, pie);
+        totalSupply = add(totalSupply, pie);
 
         daiToken.transferFrom(msg.sender, address(this), wad);
         daiJoin.join(address(this), wad);
@@ -176,24 +214,24 @@ contract Chai {
     }
 
     // wad is denominated in (1/chi) * dai
-    function exit(address src, uint wad) public {
-        require(balanceOf[src] >= wad, "chai/insufficient-balance");
-        if (src != msg.sender && allowance[src][msg.sender] != uint(-1)) {
-            require(allowance[src][msg.sender] >= wad, "chai/insufficient-allowance");
+    function exit(address src, uint256 wad) public {
+        require(balanceOf[src] >= wad, 'chai/insufficient-balance');
+        if (src != msg.sender && allowance[src][msg.sender] != uint256(-1)) {
+            require(allowance[src][msg.sender] >= wad, 'chai/insufficient-allowance');
             allowance[src][msg.sender] = sub(allowance[src][msg.sender], wad);
         }
         balanceOf[src] = sub(balanceOf[src], wad);
-        totalSupply    = sub(totalSupply, wad);
+        totalSupply = sub(totalSupply, wad);
 
-        uint chi = (now > pot.rho()) ? pot.drip() : pot.chi();
+        uint256 chi = (now > pot.rho()) ? pot.drip() : pot.chi();
         pot.exit(wad);
         daiJoin.exit(msg.sender, rmul(chi, wad));
         emit Transfer(src, address(0), wad);
     }
 
     // wad is denominated in dai
-    function draw(address src, uint wad) external {
-        uint chi = (now > pot.rho()) ? pot.drip() : pot.chi();
+    function draw(address src, uint256 wad) external {
+        uint256 chi = (now > pot.rho()) ? pot.drip() : pot.chi();
         // rounding up ensures usr gets at least wad dai
         exit(src, rdivup(wad, chi));
     }
