+++ contracts/external/maker/join.sol
@@ -18,22 +18,38 @@
 
 pragma solidity ^0.6.0;
 
-import "./lib.sol";
+import './lib.sol';
 
 interface GemLike {
-    function decimals() external view returns (uint);
-    function transfer(address,uint) external returns (bool);
-    function transferFrom(address,address,uint) external returns (bool);
+    function decimals() external view returns (uint256);
+
+    function transfer(address, uint256) external returns (bool);
+
+    function transferFrom(
+        address,
+        address,
+        uint256
+    ) external returns (bool);
 }
 
 interface DSTokenLike {
-    function mint(address,uint) external;
-    function burn(address,uint) external;
+    function mint(address, uint256) external;
+
+    function burn(address, uint256) external;
 }
 
 interface VatLike {
-    function slip(bytes32,address,int) external;
-    function move(address,address,uint) external;
+    function slip(
+        bytes32,
+        address,
+        int256
+    ) external;
+
+    function move(
+        address,
+        address,
+        uint256
+    ) external;
 }
 
 /*
@@ -62,21 +78,32 @@ interface VatLike {
 
 contract GemJoin is LibNote {
     // --- Auth ---
-    mapping (address => uint) public wards;
-    function rely(address usr) external note auth { wards[usr] = 1; }
-    function deny(address usr) external note auth { wards[usr] = 0; }
+    mapping(address => uint256) public wards;
+
+    function rely(address usr) external note auth {
+        wards[usr] = 1;
+    }
+
+    function deny(address usr) external note auth {
+        wards[usr] = 0;
+    }
+
     modifier auth {
-        require(wards[msg.sender] == 1, "GemJoin/not-authorized");
+        require(wards[msg.sender] == 1, 'GemJoin/not-authorized');
         _;
     }
 
     VatLike public vat;
     bytes32 public ilk;
     GemLike public gem;
-    uint    public dec;
-    uint    public live;  // Access Flag
+    uint256 public dec;
+    uint256 public live; // Access Flag
 
-    constructor(address vat_, bytes32 ilk_, address gem_) public {
+    constructor(
+        address vat_,
+        bytes32 ilk_,
+        address gem_
+    ) public {
         wards[msg.sender] = 1;
         live = 1;
         vat = VatLike(vat_);
@@ -84,35 +111,45 @@ contract GemJoin is LibNote {
         gem = GemLike(gem_);
         dec = gem.decimals();
     }
+
     function cage() external note auth {
         live = 0;
     }
-    function join(address usr, uint wad) external note {
-        require(live == 1, "GemJoin/not-live");
-        require(int(wad) >= 0, "GemJoin/overflow");
-        vat.slip(ilk, usr, int(wad));
-        require(gem.transferFrom(msg.sender, address(this), wad), "GemJoin/failed-transfer");
+
+    function join(address usr, uint256 wad) external note {
+        require(live == 1, 'GemJoin/not-live');
+        require(int256(wad) >= 0, 'GemJoin/overflow');
+        vat.slip(ilk, usr, int256(wad));
+        require(gem.transferFrom(msg.sender, address(this), wad), 'GemJoin/failed-transfer');
     }
-    function exit(address usr, uint wad) external note {
-        require(wad <= 2 ** 255, "GemJoin/overflow");
-        vat.slip(ilk, msg.sender, -int(wad));
-        require(gem.transfer(usr, wad), "GemJoin/failed-transfer");
+
+    function exit(address usr, uint256 wad) external note {
+        require(wad <= 2**255, 'GemJoin/overflow');
+        vat.slip(ilk, msg.sender, -int256(wad));
+        require(gem.transfer(usr, wad), 'GemJoin/failed-transfer');
     }
 }
 
 contract ETHJoin is LibNote {
     // --- Auth ---
-    mapping (address => uint) public wards;
-    function rely(address usr) external note auth { wards[usr] = 1; }
-    function deny(address usr) external note auth { wards[usr] = 0; }
+    mapping(address => uint256) public wards;
+
+    function rely(address usr) external note auth {
+        wards[usr] = 1;
+    }
+
+    function deny(address usr) external note auth {
+        wards[usr] = 0;
+    }
+
     modifier auth {
-        require(wards[msg.sender] == 1, "ETHJoin/not-authorized");
+        require(wards[msg.sender] == 1, 'ETHJoin/not-authorized');
         _;
     }
 
     VatLike public vat;
     bytes32 public ilk;
-    uint    public live;  // Access Flag
+    uint256 public live; // Access Flag
 
     constructor(address vat_, bytes32 ilk_) public {
         wards[msg.sender] = 1;
@@ -120,34 +157,44 @@ contract ETHJoin is LibNote {
         vat = VatLike(vat_);
         ilk = ilk_;
     }
+
     function cage() external note auth {
         live = 0;
     }
+
     function join(address usr) external payable note {
-        require(live == 1, "ETHJoin/not-live");
-        require(int(msg.value) >= 0, "ETHJoin/overflow");
-        vat.slip(ilk, usr, int(msg.value));
+        require(live == 1, 'ETHJoin/not-live');
+        require(int256(msg.value) >= 0, 'ETHJoin/overflow');
+        vat.slip(ilk, usr, int256(msg.value));
     }
-    function exit(address payable usr, uint wad) external note {
-        require(int(wad) >= 0, "ETHJoin/overflow");
-        vat.slip(ilk, msg.sender, -int(wad));
+
+    function exit(address payable usr, uint256 wad) external note {
+        require(int256(wad) >= 0, 'ETHJoin/overflow');
+        vat.slip(ilk, msg.sender, -int256(wad));
         usr.transfer(wad);
     }
 }
 
 contract DaiJoin is LibNote {
     // --- Auth ---
-    mapping (address => uint) public wards;
-    function rely(address usr) external note auth { wards[usr] = 1; }
-    function deny(address usr) external note auth { wards[usr] = 0; }
+    mapping(address => uint256) public wards;
+
+    function rely(address usr) external note auth {
+        wards[usr] = 1;
+    }
+
+    function deny(address usr) external note auth {
+        wards[usr] = 0;
+    }
+
     modifier auth {
-        require(wards[msg.sender] == 1, "DaiJoin/not-authorized");
+        require(wards[msg.sender] == 1, 'DaiJoin/not-authorized');
         _;
     }
 
     VatLike public vat;
     DSTokenLike public dai;
-    uint    public live;  // Access Flag
+    uint256 public live; // Access Flag
 
     constructor(address vat_, address dai_) public {
         wards[msg.sender] = 1;
@@ -155,19 +202,24 @@ contract DaiJoin is LibNote {
         vat = VatLike(vat_);
         dai = DSTokenLike(dai_);
     }
+
     function cage() external note auth {
         live = 0;
     }
-    uint constant ONE = 10 ** 27;
-    function mul(uint x, uint y) internal pure returns (uint z) {
+
+    uint256 constant ONE = 10**27;
+
+    function mul(uint256 x, uint256 y) internal pure returns (uint256 z) {
         require(y == 0 || (z = x * y) / y == x);
     }
-    function join(address usr, uint wad) external note {
+
+    function join(address usr, uint256 wad) external note {
         vat.move(address(this), usr, mul(ONE, wad));
         dai.burn(msg.sender, wad);
     }
-    function exit(address usr, uint wad) external note {
-        require(live == 1, "DaiJoin/not-live");
+
+    function exit(address usr, uint256 wad) external note {
+        require(live == 1, 'DaiJoin/not-live');
         vat.move(msg.sender, address(this), mul(ONE, wad));
         dai.mint(usr, wad);
     }
